==> 2_Working_Project_eda_data_cleaning.ipynb <==
# Selected Client:

## Charles Christensen (Seller)

#### Requirements:

    1. Invest with big returns
    2. wondering about renovation?
    3. which Neighborhood?
    4. Timing?

#### Cleaning the Data

import pandas as pd
import numpy as np

# loading data from the csv files.

df_import_sales = pd.read_csv('data/king_county_house_sales.csv')
df_import_details = pd.read_csv('data/king_county_house_details.csv')

display(df_import_details.sort_values(by='id', ascending=True).head())
display(df_import_sales.sort_values(by='house_id', ascending=True).head())

# merge data frames into a single one before cleaning
df = pd.merge(df_import_sales, df_import_details, left_on='house_id', right_on='id', how='inner')

df.rename(columns={'id_x': 'sale_id'}, inplace=True)
df = df.drop(columns=['id_y'])


display(df.head())
display(df.info())
display(df.isnull().sum())
display(df.describe())


#convert date to datetime

df["date"] = pd.to_datetime(df["date"])

#additional data columns for filtering later
df["year_sold"] = df["date"].dt.year
df["month_sold"] = df["date"].dt.month

df.info()

# as the columns: waterfront, sqft_basement & year_renovated have some null values. We'll replace them with 0

df.fillna({'waterfront': 0, 'sqft_basement':0, 'yr_renovated': 0, 'view' : 0}, inplace=True)

df.isnull().sum()

df['yr_renovated'].head()

df['yr_renovated'] = df['yr_renovated'].astype('Int64')
df['yr_renovated'] = df['yr_renovated'].astype(str)

# df['yr_renovated'].head()

def clean_year(year):
    
    if pd.isna(year) or year == 0:
        return year

    year_str = str(int(year))

    if len(year_str) == 5 and year_str.endswith('0'):
        return int(year_str[:-1])
    else:
        return year

df['yr_renovated'] = df['yr_renovated'].apply(clean_year)

display(df['yr_renovated'].head())
display(df['yr_renovated'].tail())

df.to_csv('data/filtered_after_data_cleaning.csv',index=False)
