==> 3_eda_data_analysis.ipynb <==
# Data Analysis
#### Based on the data from 2_Working_Project_eda_data_cleaning

import pandas as pd
import numpy as np

#import data frame from file

df = pd.read_csv('data/filtered_after_data_cleaning.csv')

df["house_id"].duplicated()
